+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #93 JUnit 4 support deprecation
 * #96 Stop supporting Cayenne 4.0
 * #98 CayenneTester: allow to operate unattached 
+* #99 Add a hook to run custom actions after Cayenne startup
 
 ## 2.0